Disable xfce4-screensaver
I've gone back and forth on this. The screensaver locks just fine, but it also wants to manage DPMS
which it fights with power-manager over.  I'm able to force a truce, but it just seems like the
screensaver is an extra piece that I'm using only for screen locking.

Light-locker was nice, but I'm not sure if it's a problem with the nix derivation, but it's
gsettings schemas was not visible to power-manager unless I manually changed shell variables.  Since
power-manager couldn't see the schemas it would not add the "Security" tab for configuring
light-locker, which was not really necessary, but just really annoyed me.

Finally I realized I should just be able to use lightdm to lock, so I configured it as the
LockCommand and disabled the screensaver and everything is good, for now.