Recognize TOLISTEN, TOREAD, TOWATCH as inactive
My inactive check is a simple set intersection, which does not take into account any tag hierarchy.
I spent a bit of time (too much) trying to figure out how to respect the tag hierarchy, and
eventually--in frustration--decided the easiest thing to do was the hackiest and just add the tags
to the inactive list.

I'll have to keep this list updated if things change, which is error prone, but also not disastrous.

I may try again on the tag hierarchy, but I need to give it time.